feat: refactor trade handling to use Quantity value object and update response types

+++ src/Scheduled.Stock.Purchase.Api/Features/Trades/CreateTrade/CreateTradeResponse.cs
@@ -6,7 +6,7 @@ internal sealed class CreateTradeResponse
 {
     public Guid Id { get; init; }
     public string Ticker { get; init; } = default!;
-    public decimal Quantity { get; init; }
+    public int Quantity { get; init; }
     public decimal Price { get; init; }
     public TradeType Type { get; init; }
     public DateTime ExecutedAt { get; init; }
@@ -14,7 +14,7 @@ internal sealed class CreateTradeResponse
     public CreateTradeResponse(
         Guid id,
         string ticker,
-        decimal quantity,
+        int quantity,
         decimal price,
         TradeType type,
         DateTime executedAt
